# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusSathi API for locally stored notes and keyword-based study tools.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Summary {
  topic: string;
  bullets: string[];
  formulas: string[];
  definitions: string[];
  sourceWordCount: number;
}

export type NoteSubject = typeof NoteSubject[keyof typeof NoteSubject];


export const NoteSubject = {
  DSA: 'DSA',
  DBMS: 'DBMS',
  OS: 'OS',
  CN: 'CN',
} as const;

export interface Note {
  id: string;
  ownerId: string;
  title: string;
  subject: NoteSubject;
  filename: string;
  fileSize: number;
  uploadedAt: string;
  isShared: boolean;
  summary: Summary;
}

export interface Dashboard {
  totalNotes: number;
  sharedNotes: number;
  totalWords: number;
  hoursSaved: number;
}

export interface DashboardInput {
  ownerId: string;
}

export type UploadNoteInputSubject = typeof UploadNoteInputSubject[keyof typeof UploadNoteInputSubject];


export const UploadNoteInputSubject = {
  DSA: 'DSA',
  DBMS: 'DBMS',
  OS: 'OS',
  CN: 'CN',
} as const;

export interface UploadNoteInput {
  ownerId: string;
  title: string;
  subject: UploadNoteInputSubject;
  file: Blob;
}

export interface MineNotesInput {
  ownerId: string;
}

export interface ShareNoteInput {
  ownerId: string;
  noteId: string;
  isShared: boolean;
}

export interface QuestionInput {
  ownerId: string;
  noteId: string;
  /** @minLength 1 */
  question: string;
}

export interface DeleteNoteInput {
  ownerId: string;
  noteId: string;
}

export interface DownloadNoteInput {
  ownerId: string;
  noteId: string;
}

export interface AnswerResult {
  answer: string;
  found: boolean;
  matchedKeywords: string[];
}