# lib/api-zod/src/generated/types/deleteNoteInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusSathi API for locally stored notes and keyword-based study tools.
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteNoteInput {
  ownerId: string;
  noteId: string;
}