# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusSathi API for locally stored notes and keyword-based study tools.
 * OpenAPI spec version: 0.1.0
 */
import type { NoteSubject } from './noteSubject';
import type { Summary } from './summary';

export interface Note {
  id: string;
  ownerId: string;
  title: string;
  subject: NoteSubject;
  filename: string;
  fileSize: number;
  uploadedAt: string;
  isShared: boolean;
  summary: Summary;
}