# lib/api-zod/src/generated/types/questionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusSathi API for locally stored notes and keyword-based study tools.
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionInput {
  ownerId: string;
  noteId: string;
  /** @minLength 1 */
  question: string;
}