# lib/api-zod/src/generated/types/uploadNoteInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusSathi API for locally stored notes and keyword-based study tools.
 * OpenAPI spec version: 0.1.0
 */
import type { UploadNoteInputSubject } from './uploadNoteInputSubject';

export interface UploadNoteInput {
  ownerId: string;
  title: string;
  subject: UploadNoteInputSubject;
  file: Blob;
}